Move connection_candidate
