+++ src/main/mod.rs
@@ -21,7 +21,8 @@ use std::sync::{Arc, Mutex};
 pub use self::active_connection::{ActiveConnection, INACTIVITY_TIMEOUT_MS};
 pub use self::bootstrap::Bootstrap;
 pub use self::config_handler::Config;
-pub use self::connect::{Connect, ConnectionCandidate};
+pub use self::connect::Connect;
+pub use self::connection_candidate::ConnectionCandidate;
 pub use self::connection_listener::ConnectionListener;
 pub use self::event::Event;
 pub use self::error::CrustError;
@@ -35,6 +36,7 @@ mod active_connection;
 mod bootstrap;
 mod config_handler;
 mod connect;
+mod connection_candidate;
 mod connection_listener;
 mod event;
 mod error;
